// src/api/generated/model/storeUpdateRequest.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * 뭐든배달 API 명세서
 * 뭐든배달 API 명세서입니다
 * OpenAPI spec version: v1
 */

export interface StoreUpdateRequest {
  storeCategoryId?: number;
  name?: string;
  description?: string;
  roadAddr?: string;
  lat?: number;
  lng?: number;
  /**
   * @minLength 1
   * @maxLength 256
   */
  imageUrl?: string;
}